Better module installation support with progress reporting
