+++ src/main/java/org/terasology/web/resources/modules/AvailableModulesResource.java
@@ -15,29 +15,21 @@
  */
 package org.terasology.web.resources.modules;
 
-import org.terasology.engine.module.DependencyResolutionFailedException;
 import org.terasology.engine.module.ModuleManager;
 import org.terasology.i18n.I18nMap;
 import org.terasology.module.Module;
 import org.terasology.module.ModuleMetadata;
-import org.terasology.naming.Name;
 import org.terasology.network.Client;
 import org.terasology.registry.In;
-import org.terasology.utilities.download.MultiFileTransferProgressListener;
-import org.terasology.web.ServerAdminsManager;
-import org.terasology.web.io.ActionResult;
 import org.terasology.web.resources.ReadableResource;
 import org.terasology.web.resources.ResourceAccessException;
-import org.terasology.web.resources.WritableResource;
 import org.terasology.world.generator.internal.WorldGeneratorManager;
 
 import java.util.Comparator;
-import java.util.Set;
-import java.util.concurrent.Callable;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-public class AvailableModulesResource implements ReadableResource<AvailableModulesData>, WritableResource<Name[]> {
+public class AvailableModulesResource implements ReadableResource<AvailableModulesData> {
 
     @In
     private ModuleManager moduleManager;
@@ -55,38 +47,5 @@ public class AvailableModulesResource implements ReadableResource<AvailableModul
         Stream<ModuleMetadata> modules = moduleManager.getRegistry().stream().map(Module::getMetadata)
                 .sorted(Comparator.comparing(ModuleMetadata::getDisplayName, Comparator.comparing(I18nMap::value)));
         return new AvailableModulesData(modules.collect(Collectors.toList()), worldGeneratorManager.getWorldGenerators());
-    }
-
-    @Override
-    public Class<Name[]> getDataType() {
-        return Name[].class;
-    }
-
-    @Override
-    public void write(Client requestingClient, Name[] data) throws ResourceAccessException {
-        ServerAdminsManager.checkClientIsServerAdmin(requestingClient.getId());
-        executeCallable(moduleManager.getInstallManager().updateRemoteRegistry());
-        Set<Module> allModules;
-        try {
-            allModules = moduleManager.getInstallManager().getAllModulesToDownloadFor(data);
-        } catch (DependencyResolutionFailedException ex) {
-            throw new ResourceAccessException(new ActionResult(ActionResult.Status.GENERIC_ERROR, ex.getMessage()));
-        }
-        executeCallable(moduleManager.getInstallManager().createInstaller(allModules, new MultiFileTransferProgressListener() {
-            @Override
-            public void onSizeMetadataProgress(int index, int totalUrls) {
-            }
-            @Override
-            public void onDownloadProgress(long totalTransferredBytes, long totalBytes, int completedFiles, int nFiles) {
-            }
-        }));
-    }
-
-    private void executeCallable(Callable process) throws ResourceAccessException {
-        try {
-            process.call();
-        } catch (Exception ex) {
-            throw new ResourceAccessException(new ActionResult(ActionResult.Status.GENERIC_ERROR, ex.getMessage()));
-        }
     }
 }
